+++ src/snowflake/cli/plugins/nativeapp/commands.py
@@ -24,13 +24,19 @@ from snowflake.cli.api.commands.decorators import (
     with_project_definition,
 )
 from snowflake.cli.api.commands.snow_typer import SnowTyperFactory
+from snowflake.cli.api.output.formats import OutputFormat
 from snowflake.cli.api.output.types import (
     CollectionResult,
     CommandResult,
     MessageResult,
+    ObjectResult,
 )
 from snowflake.cli.api.secure_path import SecurePath
-from snowflake.cli.plugins.nativeapp.common_flags import ForceOption, InteractiveOption
+from snowflake.cli.plugins.nativeapp.common_flags import (
+    ForceOption,
+    InteractiveOption,
+    ValidateOption,
+)
 from snowflake.cli.plugins.nativeapp.init import (
     OFFICIAL_TEMPLATES_GITHUB_URL,
     nativeapp_init,
@@ -82,7 +88,7 @@ def app_init(
     ),
     template: str = typer.Option(
         None,
-        help="A specific template name within the template repo to use as template for the Native Apps project. Example: Default is basic if `--template-repo` is https://github.com/snowflakedb/native-apps-templates.git, and None if any other --template-repo is specified.",
+        help="A specific template name within the template repo to use as template for the Snowflake Native App project. Example: Default is basic if `--template-repo` is https://github.com/snowflakedb/native-apps-templates.git, and None if any other --template-repo is specified.",
     ),
     **options,
 ) -> CommandResult:
@@ -175,6 +181,7 @@ def app_run(
     ),
     interactive: bool = InteractiveOption,
     force: Optional[bool] = ForceOption,
+    validate: bool = ValidateOption,
     **options,
 ) -> CommandResult:
     """
@@ -203,6 +210,7 @@ def app_run(
         patch=patch,
         from_release_directive=from_release_directive,
         is_interactive=is_interactive,
+        validate=validate,
     )
     return MessageResult(
         f"Your application object ({processor.app_name}) is now available:\n"
@@ -273,6 +281,7 @@ def app_deploy(
         show_default=False,
         help=f"""Paths, relative to the the project root, of files you want to upload to a stage. The paths must match one of the artifacts src pattern entries in snowflake.yml. If unspecified, the command syncs all local changes to the stage.""",
     ),
+    validate: bool = ValidateOption,
     **options,
 ) -> CommandResult:
     """
@@ -300,8 +309,26 @@ def app_deploy(
         prune=prune,
         recursive=recursive,
         local_paths_to_sync=files,
+        validate=validate,
     )
 
     return MessageResult(
         f"Deployed successfully. Application package and stage are up-to-date."
     )
+
+
+@app.command("validate", requires_connection=True)
+@with_project_definition("native_app")
+def app_validate(**options):
+    """
+    Validates a deployed Snowflake Native App's setup script.
+    """
+    manager = NativeAppManager(
+        project_definition=cli_context.project_definition,
+        project_root=cli_context.project_root,
+    )
+    if cli_context.output_format == OutputFormat.JSON:
+        return ObjectResult(manager.get_validation_result(use_scratch_stage=True))
+
+    manager.validate(use_scratch_stage=True)
+    return MessageResult("Snowflake Native App validation succeeded.")